Moved cameraclasses to a separat folder
