+++ StortSpelProjekt/Engine/src/ECS/Components/InputComponent.cpp
@@ -1,7 +1,7 @@
 #include "InputComponent.h"
 #include "../Events/EventBus.h"
 #include "../Entity.h"
-#include "../Renderer/PerspectiveCamera.h"
+#include "../Renderer/Camera/PerspectiveCamera.h"
 #include "../Renderer/Transform.h"
 
 component::InputComponent::InputComponent(Entity* parent)
