+++ StortSpelProjekt/Engine/src/ECS/Components/Lights/SpotLightComponent.cpp
@@ -4,7 +4,7 @@
 
 // Renderer
 #include "../Renderer/Transform.h"
-#include "../Renderer/BaseCamera.h"
+#include "../Renderer/Camera/BaseCamera.h"
 #include "../Renderer/Renderer.h"
 #include "../Misc/Option.h"
