+++ StortSpelProjekt/Engine/src/Renderer/DX12Tasks/OutliningRenderTask.cpp
@@ -10,7 +10,7 @@
 #include "../SwapChain.h"
 #include "../Renderer/Transform.h"
 #include "../Renderer/Mesh.h"
-#include "../Renderer/BaseCamera.h"
+#include "../Renderer/Camera/BaseCamera.h"
 #include "../GPUMemory/RenderTargetView.h"
 #include "../GPUMemory/DepthStencil.h"
 #include "../GPUMemory/DepthStencilView.h"
