+++ StortSpelProjekt/Engine/src/Renderer/DX12Tasks/SkyboxRenderTask.cpp
@@ -11,7 +11,7 @@
 #include "../PipelineState.h"
 #include "../Renderer/Transform.h"
 #include "../Renderer/Mesh.h"
-#include "../BaseCamera.h"
+#include "../Camera/BaseCamera.h"
 #include "../SwapChain.h"
 #include "../GPUMemory/RenderTarget.h"
 #include "../GPUMemory/RenderTargetView.h"
